add FeaturesConnected component for dynamic name.
