+++ packages/apolloschurchapp/src/ui/Onboarding/slides/Features/getUserFirstName.js
@@ -1,0 +1,12 @@
+import gql from 'graphql-tag';
+
+export default gql`
+  query getUserFirstName {
+    currentUser {
+      id
+      profile {
+        firstName
+      }
+    }
+  }
+`;
