+++ packages/apolloschurchapp/src/ui/Onboarding/slides/Features/index.js
@@ -1,40 +1,4 @@
-import React, { memo } from 'react';
-import PropTypes from 'prop-types';
+import Features from './Features';
 
-import Slide, { SlideContent } from '../../Slide';
-
-const Features = memo(
-  ({ firstName, description, BackgroundComponent, ...props }) => (
-    <Slide {...props}>
-      {BackgroundComponent}
-      <SlideContent title={`Hey ${firstName}!`} description={description} />
-    </Slide>
-  )
-);
-
-Features.displayName = 'Features';
-
-Features.propTypes = {
-  /* The `Swiper` component used in `<onBoarding>` looks for and hijacks the title prop of it's
-   * children. Thus we have to use more unique name.
-   */
-  firstName: PropTypes.string,
-  description: PropTypes.string,
-  /* Recommended usage:
-   * - `Image` (react-native)
-   * - `GradientOverlayImage` (@apollosproject/ui-kit) for increased readability
-   * - `Video` (react-native-video) because moving pictures!
-   */
-  BackgroundComponent: PropTypes.oneOfType([
-    PropTypes.arrayOf(PropTypes.node),
-    PropTypes.node,
-  ]),
-};
-
-Features.defaultProps = {
-  firstName: 'friend',
-  description:
-    "We'd like to help personalize your mobile experience so we can help you with every step on your journey.",
-};
-
+export FeaturesConnected from './FeaturesConnected';
 export default Features;
